Fix: (`typescript-vm`) - Simplify SWC handling for source maps in typescript-vm
Removed the dependency on node's inspector module, as well as the condition to check if the debugger is present to decide inlining source maps. This dependency and condition became unnecessary as we modified the way SWC handles source maps. Now, the SWC source map generation is always happening, just ensuring it generates inline source maps for better compatibility.

Furthermore, refactored the code for better readability and encapsulation. In the process of this change, 'uncaughtException' event handling has been removed as it was not helping in handling SWC's inability to deal with valid source maps. Now, the source map correction is happening directly inside the 'retrieveSourceMap' callback.

Lastly, clarified the options passed into the 'swc.transform' call, and also the return statement is changed to include the base64 encoded source map. Overall, this change makes our source map handling independent of the execution environment and provides more robust error handling.