Fix: (`tsconfig-to-swc`) - Fix importInterop assignment
Corrected the ternary operator used to assign the 'importInterop' value in zenflux-tsconfig-to-swc. Instead of relying on the nullish coalescing (??) operator which was potentially causing incorrect "swc" assignments, it now correctly assigns "swc" or "none" based on the 'esModuleInterop' value. This change should prevent unexpected behavior when esModuleInterop is undefined or null.