Tweak: (`cli`) - Add new Interface IZConfigInMulti to config definitions
A new interface `IZConfigInMulti` is added to the config definitions in `config.ts`. This interface extends both `IConfigArgsForEachFormat` and `IConfigArgsGeneral`. The purpose of this change is to allow the use of `IZConfigInMulti` in key-value pairs for `IZConfig` type definition as well, broadening the spectrum for usage of config settings in our tool and making the configuration more dynamic.