Tweak: (`cli`) - Improve error handling in zenflux-cli build process
An update has been made to the build.ts file in the zenflux-cli package to enhance error handling during the build process. Firstly, a new onLog event handler is introduced for the Rollup configuration. This handler will output log messages from Rollup and also attach the project path to these messages. Secondly, the logic for handling errors during the rollupBuildInternal task has been encapsulated in a try-catch block. When an error occurs, a detailed error message, including the stack trace and the configuration path, is sent to the parent process. Finally, the handling of internal worker errors was updated. Now, in case of an error, the failure details are not thrown as exceptions but instead are logged to the console, and the process exits with a failure status.