New: (`cli`) - Add function to create TypeScript declaration
A new function, zTSCreateDeclaration, has been added in the typescript.ts file. This function creates a TypeScript declaration based on the provided tsConfig. It extends its options to include necessary parameters for declaration creation. This will be useful for automating the process of declaration file generation. Old declaration files are also cleaned up in the process.