Tweak: (`cli`) - Implement multithreading and diagnostics for build command
In this update, we've integrated a multithreading mechanism for handling configurations in the build process. The aim is to increment the efficiency for large projects.
We also added a pre-diagnostic tool with a 'haltOnError' flag which lets you decide if the process should be killed when a diagnostic error occurs. Also, for each configuration, a declaration file is created after the build.
A '--haltOnDiagnosticError' command was added, which halts process execution in case of a typescript diagnostic error. The feature makes debugging more effective by pinpointing the exact configuration throwing the error.