Tweak: (`typescript-vm`) - Add 'refererUrl' to ImportMeta interface
This commit introduces a new 'refererUrl' attribute to the ImportMeta interface, following TypeScript definition patterns. It helps in keeping the reference of the module that the current one is being imported into. The loaders.js file was altered to handle this addition while maintaining the module loading function. Changes were also made to the index.js file for accurate module referencing.