Fix: (`zenflux-jest`) - Replace console with ConsoleManager and update cache configuration
In this commit, all the instances of console in the code were replaced with ConsoleManager for standardization and improved message handling. Additionally, the cache configuration was altered to account for inspector URL and a new property for testTimeout was added. Logic was introduced to allow tests to run by exact name when the gate pragma is not specified.