Tweak: (`zenflux-cli`) - Improve Typescript and Rollup handling
The code changes improved the behavior and error handling of Typescript and Rollup build tasks within the CLI. Notably, the handling of dependencies in multi-threaded builds was advanced. Unnecessary/unused import was also removed in "common.ts". Other updates include improvements to console logging and error handling for clarity and better debugging. The 'zTSPreDiagnostics' function was further refactored for improved accuracy.