New: (`github-workflows`) - Implement main workflow and update react packages workflow
Added a new `main.yml` for the primary GitHub Actions workflow that is executed on every push, or pull request. This main workflow contains two jobs: one for self-hosted runners and another for GitHub-hosted runners (fallback when self-hosted job fails). Adjustments were also made to the `react-packages.yml` workflow, allowing it to be called within the main workflow with a dynamic runner input. The GITHUB_TOKEN is passed as a secret to these workflows.