Update eslint dependency and bump package version (#92)
* Update eslint dependency and bump package version

Upgraded eslint from ^9.10.0 to ^9.23.0 to incorporate the latest fixes and features. Incremented the package version to 0.1.3 to reflect the update and ensure proper versioning.

* Refactor: Update ESLint type definitions for compatibility

Modified the `default-config.d.ts` to enhance the `ESLintTSLintCompatible` type by allowing a more flexible `languageOptions` structure. Updated the return type of `zLintDefaultConfig` to an array for better handling of multiple configurations.

Adjusted `main.d.ts` to import the new type and updated function return types for `zLintGetBaseConfig` and `zLintGetConfig` to ensure consistency across the codebase.

* Exports default configuration

Exports the default configuration from the main module, simplifying its usage and improving code modularity.

Updates version to 0.1.6.